docs/requirements.txt
sphinx-codeautolink==0.10.0
sphinx-toolbox==2.16.1
sphinx-rtd-theme==1.0.0
sphinx-rtd-dark-mode==1.2.4
sphinx-copybutton==0.5.0
sphinx-tabs==3.2.0
sphinx-prompt==1.5.0
sphinx-hoverxref==1.0.1
